refactor: add lombok
