+++ src/main/java/com/innowise/salesAndCustomerAnalytics/entities/Category.java
@@ -1,4 +1,4 @@
-package main;
+package com.innowise.salesAndCustomerAnalytics.entities;
 
 public enum Category {
     ELECTRONICS, CLOTHING, BOOKS, HOME, BEAUTY, TOYS
