+++ src/main/java/com/innowise/salesAndCustomerAnalytics/entities/Order.java
@@ -1,41 +1,16 @@
-package main;
+package com.innowise.salesAndCustomerAnalytics.entities;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Data;
 
+@Data
+@AllArgsConstructor
 public class Order {
     private String orderId;
     private LocalDateTime orderDate;
     private Customer customer;
     private List<OrderItem> items;
     private OrderStatus status;
-
-    public Order(String orderId, LocalDateTime orderDate, Customer customer, List<OrderItem> items,
-                 OrderStatus status) {
-        this.orderId = orderId;
-        this.orderDate = orderDate;
-        this.customer = customer;
-        this.items = items;
-        this.status = status;
-    }
-
-    public String getOrderId() {
-        return orderId;
-    }
-
-    public LocalDateTime getOrderDate() {
-        return orderDate;
-    }
-
-    public Customer getCustomer() {
-        return customer;
-    }
-
-    public List<OrderItem> getItems() {
-        return items;
-    }
-
-    public OrderStatus getStatus() {
-        return status;
-    }
 }
