+++ src/main/java/com/innowise/salesAndCustomerAnalytics/entities/OrderItem.java
@@ -1,31 +1,13 @@
-package main;
+package com.innowise.salesAndCustomerAnalytics.entities;
 
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
 public class OrderItem {
     private String productName;
     private int quantity;
     private double price;
     private Category category;
-
-    public OrderItem(String productName, int quantity, double price, Category category) {
-        this.productName = productName;
-        this.quantity = quantity;
-        this.price = price;
-        this.category = category;
-    }
-
-    public String getProductName() {
-        return productName;
-    }
-
-    public int getQuantity() {
-        return quantity;
-    }
-
-    public double getPrice() {
-        return price;
-    }
-
-    public Category getCategory() {
-        return category;
-    }
 }
