+++ src/main/java/com/innowise/salesAndCustomerAnalytics/entities/OrderStatus.java
@@ -1,4 +1,4 @@
-package main;
+package com.innowise.salesAndCustomerAnalytics.entities;
 
 public enum OrderStatus {
     NEW, PROCESSING, SHIPPED, DELIVERED, CANCELLED
